restructured landing page with components
